break a bunch of utility functions into their own file
